requirements.txt
django==5.2.4
django-environ==0.12.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
psycopg2==2.9.10
uvicorn==0.35.0
pandas==2.3.1
django-storages[google]==1.14.6